Change TG recovery install script filename [2/2]
This is part 2/2 to maintain backwards compatibility with CWM's
verify_root_and_recovery() function. CWM checks if install-recovery.sh
exists and has an executable flag set, then offers to disable the script
for the user. CM now controls this with the persist.sys.recovery_update
property which is configurable via settings, so we don't need to
double-check this.

This changes the name of the recovery install script to
install-tg-recovery.sh.

Change-Id: I275dd358b46c626dfcb8fe02c583a308d5a89c56